Thank you so much for this project. ❤️

## 📋 Issue Summary

I've discovered a **probability distribution bug** in PocketBase's regex-based random string generation system that affects the randomness and uniqueness of generated identifiers.

## 🔍 Problem Description

When using character classes in regex patterns for ID generation, the current implementation creates a **severe bias** toward smaller character ranges. This is particularly problematic when using URL-safe base64 patterns like `[A-Za-z0-9\-_]{30}`, where single characters (`-` and `_`) appear far more frequently than they should.

The issue also affects the default pattern `[a-z0-9]{15}`, creating uneven distribution between letters and numbers.

## 🎯 Examples of the Problem

Here are some examples of generated strings using `[A-Za-z0-9\-_]{30}` showing the excessive occurrence of `-` and `_`:

```
Vw39d_h776XXabv-NJ_---_2-_Y__J
--i_6yLb_3-io54_8NP-5Vp_I-JZb6
i_UoCw-2---W-__w35_3RvS__-5-M-
--sLJdC-_lH-xD_9--9fx-_Xb4_-m-
-M15_qe_R3C_fVp7M--_-715-mDUeo
f-b-MImP_xhg4Y-O--_WrdGK-_-_2s
lnoE--_yN-m0_b2_L7timLg_FzaRhX
_S-S_v-q4Z-H____q0_-i-____I__V
-_6-___2PvL_-K__tNg_T__B-Oqg50
1O7NH__r2G_aI-9-1--_ok--__v-62
```

> **Notice**: The strings are heavily biased toward `-` and `_` characters, creating poor randomness distribution.

After this fix the generated strings look like this:

```
AJAW5tt737M-tDCdVjjpQMfiwlvD9B
F5mGcR4xuGU38EjFb-L5GGxCP0iPkd
E7zoxVWKAGf29agVoXQdUNY2VH_CLw
Ue8SfAfM9ASJ7KPR3y6PQoGcxky2Y_
8hnPS6nqv7tAhl_fpsgoHiXTbGKq1F
VCt9AW2YdlHA294ee8MRI0ALw64UEQ
AOLraN8t6UvyB6acRzHavlDXpYi3aZ
XnYr3Euap_6llHaeEnQXnlTqudQkZn
jCsYcKn1fepprFlwFzdwInvP5t36rs
KfAaZcDEwq_E-jDat3Eu__cV85Z9Ul
```

## ⚡ Cause

The current implementation in `tools/security/random_by_regex.go` processes character classes incorrectly:

```go
func randomRuneFromPairs(pairs []rune) (rune, error) {
    idx, err := randomNumber(len(pairs) / 2)
    if err != nil {
        return 0, err
    }
    return randomRuneFromRange(pairs[idx*2], pairs[idx*2+1])
}
```

### Splitting into Ranges

For the pattern `[A-Za-z0-9\-_]{30}`, the regex compiler creates these range pairs:

| Range | Characters | Count | Current Probability | Expected Probability |
| ----- | ---------- | ----- | ------------------- | -------------------- |
| `A-Z` | A,B,C...Z  | 26    | **20%**             | ~40.6%               |
| `a-z` | a,b,c...z  | 26    | **20%**             | ~40.6%               |
| `0-9` | 0,1,2...9  | 10    | **20%**             | ~15.6%               |
| `-`   | -          | 1     | **20%**             | ~1.6%                |
| `_`   | \_         | 1     | **20%**             | ~1.6%                |

### The Problem

❌ **Current approach**: Randomly selects a _range_ first (each gets 20%), then picks a character from that range  
✅ **Better approach**: Each individual character should have equal probability (~1.56% for 64 total characters)

**Result**: Single characters get **12.8x higher probability** than they should!

### Additional Context

The input pairs array for the base64 pattern becomes:  
`['-', '-', '0', '9', 'A', 'Z', '_', '_', 'a', 'z']`

Even when writing the regex explicitly as:  
`[ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789\-_]{30}`  
The compiler still optimizes it to ranges, resulting in the same biased behavior.

---

## 🛠️ Proposed Solution

I've developed a corrected implementation that uses **weighted random selection** to ensure uniform character distribution:

```go
func randomRuneFromPairs(pairs []rune) (rune, error) {
	if len(pairs)%2 != 0 {
		return 0, fmt.Errorf("invalid pairs slice: odd number of elements")
	}

	// Pre-calculate the cumulative size of all ranges to make the selection process more efficient.
	cumulativeSizes := make([]int, len(pairs)/2)
	totalRunes := 0
	for i := 0; i < len(pairs); i += 2 {
		start, end := pairs[i], pairs[i+1]
		if start > end {
			return 0, fmt.Errorf("invalid range: start '%c' > end '%c'", start, end)
		}
		totalRunes += int(end - start + 1)
		cumulativeSizes[i/2] = totalRunes
	}

	if totalRunes == 0 {
		return 0, errors.New("no runes to choose from")
	}

	// Select a random number in the range of total runes.
	runeNumber, err := randomNumber(totalRunes)
	if err != nil {
		return 0, fmt.Errorf("failed to generate random number: %w", err)
	}

	// Find which range the selected number falls into using the pre-calculated cumulative sizes.
	for i, size := range cumulativeSizes {
		if runeNumber < size {
			startRune := pairs[i*2]
			previousSize := 0
			if i > 0 {
				previousSize = cumulativeSizes[i-1]
			}
			return startRune + rune(runeNumber-previousSize), nil
		}
	}

	// This part should be unreachable if the logic is correct.
	// It indicates a bug in this function or in randomNumber.
	panic("unreachable: failed to find a rune")
}
```

🎯 This assures a correct probability distribution: Each character gets equal probability.

---

## 📊 Performance Benchmarks

I conducted tests and benchmarks to validate both correctness and performance improvements:

### Benchmark Results

| Implementation      | Pattern              | Iterations | Time/op  | Memory/op | Allocs/op |
| ------------------- | -------------------- | ---------- | -------- | --------- | --------- |
| **Current (Buggy)** | `[a-z0-9]{15}`       | 361,990    | 3,236 ns | 1,840 B   | 97        |
| **Fixed**           | `[a-z0-9]{15}`       | 477,039    | 2,433 ns | 1,360 B   | 67        |
| **Current (Buggy)** | `[A-Za-z0-9\-_]{30}` | 187,546    | 6,181 ns | 3,312 B   | 178       |
| **Fixed**           | `[A-Za-z0-9\-_]{30}` | 271,593    | 4,237 ns | 3,408 B   | 130       |

---

## 🧪 Statistical Validation

I ran extensive statistical tests with 500,000 iterations to verify the probability distribution fixes:

### 🎯 Key Results Summary

| Pattern              | Implementation  | Standard Deviation | Status                  |
| -------------------- | --------------- | ------------------ | ----------------------- |
| `[a-z0-9]{15}`       | Current (Buggy) | **1.40%**          | ❌ Biased               |
| `[a-z0-9]{15}`       | Fixed           | **0.01%**          | ✅ Uniform              |
| `[A-Za-z0-9\-_]{30}` | Current (Buggy) | **3.37%**          | ❌ Severely Biased      |
| `[A-Za-z0-9\-_]{30}` | Fixed           | **0.00%**          | ✅ Perfect Distribution |

### Detailed Test Results

#### Lower Alpha-Numerical Pattern: `[a-z0-9]{15}` _(500k iterations)_

<details>
<summary>Current Implementation standard deviation: <strong>1.40%</strong></summary>

- `0`: Count: 373935, Percentage: 4.99%
- `1`: Count: 375151, Percentage: 5.00%
- `2`: Count: 375199, Percentage: 5.00%
- `3`: Count: 375345, Percentage: 5.00%
- `4`: Count: 375677, Percentage: 5.01%
- `5`: Count: 375172, Percentage: 5.00%
- `6`: Count: 375182, Percentage: 5.00%
- `7`: Count: 376248, Percentage: 5.02%
- `8`: Count: 374385, Percentage: 4.99%
- `9`: Count: 375316, Percentage: 5.00%
- `a`: Count: 144097, Percentage: 1.92%
- `b`: Count: 144382, Percentage: 1.93%
- `c`: Count: 143071, Percentage: 1.91%
- `d`: Count: 144743, Percentage: 1.93%
- `e`: Count: 143769, Percentage: 1.92%
- `f`: Count: 144354, Percentage: 1.92%
- `g`: Count: 144086, Percentage: 1.92%
- `h`: Count: 144624, Percentage: 1.93%
- `i`: Count: 144393, Percentage: 1.93%
- `j`: Count: 144134, Percentage: 1.92%
- `k`: Count: 144344, Percentage: 1.92%
- `l`: Count: 143677, Percentage: 1.92%
- `m`: Count: 143942, Percentage: 1.92%
- `n`: Count: 144387, Percentage: 1.93%
- `o`: Count: 144265, Percentage: 1.92%
- `p`: Count: 143811, Percentage: 1.92%
- `q`: Count: 144057, Percentage: 1.92%
- `r`: Count: 144318, Percentage: 1.92%
- `s`: Count: 144086, Percentage: 1.92%
- `t`: Count: 143582, Percentage: 1.91%
- `u`: Count: 144125, Percentage: 1.92%
- `v`: Count: 144619, Percentage: 1.93%
- `w`: Count: 144542, Percentage: 1.93%
- `x`: Count: 144634, Percentage: 1.93%
- `y`: Count: 144549, Percentage: 1.93%
- `z`: Count: 143799, Percentage: 1.92%

</details>

<details>
<summary>Fixed Implementation standard deviation: <strong>0.01%</strong></summary>

- `0`: Count: 207817, Percentage: 2.77%
- `1`: Count: 208467, Percentage: 2.78%
- `2`: Count: 208336, Percentage: 2.78%
- `3`: Count: 207442, Percentage: 2.77%
- `4`: Count: 207867, Percentage: 2.77%
- `5`: Count: 207851, Percentage: 2.77%
- `6`: Count: 208950, Percentage: 2.79%
- `7`: Count: 208687, Percentage: 2.78%
- `8`: Count: 208469, Percentage: 2.78%
- `9`: Count: 208209, Percentage: 2.78%
- `a`: Count: 208805, Percentage: 2.78%
- `b`: Count: 207595, Percentage: 2.77%
- `c`: Count: 208346, Percentage: 2.78%
- `d`: Count: 208771, Percentage: 2.78%
- `e`: Count: 208408, Percentage: 2.78%
- `f`: Count: 208171, Percentage: 2.78%
- `g`: Count: 208835, Percentage: 2.78%
- `h`: Count: 208600, Percentage: 2.78%
- `i`: Count: 208525, Percentage: 2.78%
- `j`: Count: 207299, Percentage: 2.76%
- `k`: Count: 208349, Percentage: 2.78%
- `l`: Count: 208224, Percentage: 2.78%
- `m`: Count: 208648, Percentage: 2.78%
- `n`: Count: 208462, Percentage: 2.78%
- `o`: Count: 207665, Percentage: 2.77%
- `p`: Count: 208658, Percentage: 2.78%
- `q`: Count: 208817, Percentage: 2.78%
- `r`: Count: 208350, Percentage: 2.78%
- `s`: Count: 208447, Percentage: 2.78%
- `t`: Count: 208835, Percentage: 2.78%
- `u`: Count: 208583, Percentage: 2.78%
- `v`: Count: 208496, Percentage: 2.78%
- `w`: Count: 209050, Percentage: 2.79%
- `x`: Count: 207390, Percentage: 2.77%
- `y`: Count: 208524, Percentage: 2.78%
- `z`: Count: 208052, Percentage: 2.77%

</details>

#### Alpha-Numerical Pattern: `[A-Za-z0-9]{15}` _(500k iterations)_

<details>
<summary>Current Implementation standard deviation: <strong>0.76%</strong></summary>

- `0`: Count: 250800, Percentage: 3.34%
- `1`: Count: 249670, Percentage: 3.33%
- `2`: Count: 249506, Percentage: 3.33%
- `3`: Count: 250264, Percentage: 3.34%
- `4`: Count: 250538, Percentage: 3.34%
- `5`: Count: 250189, Percentage: 3.34%
- `6`: Count: 249382, Percentage: 3.33%
- `7`: Count: 249545, Percentage: 3.33%
- `8`: Count: 250473, Percentage: 3.34%
- `9`: Count: 250856, Percentage: 3.34%
- `A`: Count: 96462, Percentage: 1.29%
- `B`: Count: 96542, Percentage: 1.29%
- `C`: Count: 96366, Percentage: 1.28%
- `D`: Count: 95718, Percentage: 1.28%
- `E`: Count: 95807, Percentage: 1.28%
- `F`: Count: 96264, Percentage: 1.28%
- `G`: Count: 96948, Percentage: 1.29%
- `H`: Count: 95795, Percentage: 1.28%
- `I`: Count: 96333, Percentage: 1.28%
- `J`: Count: 96121, Percentage: 1.28%
- `K`: Count: 95776, Percentage: 1.28%
- `L`: Count: 95860, Percentage: 1.28%
- `M`: Count: 96221, Percentage: 1.28%
- `N`: Count: 96520, Percentage: 1.29%
- `O`: Count: 96238, Percentage: 1.28%
- `P`: Count: 96060, Percentage: 1.28%
- `Q`: Count: 96221, Percentage: 1.28%
- `R`: Count: 96354, Percentage: 1.28%
- `S`: Count: 96226, Percentage: 1.28%
- `T`: Count: 95666, Percentage: 1.28%
- `U`: Count: 95872, Percentage: 1.28%
- `V`: Count: 95690, Percentage: 1.28%
- `W`: Count: 95574, Percentage: 1.27%
- `X`: Count: 96255, Percentage: 1.28%
- `Y`: Count: 96027, Percentage: 1.28%
- `Z`: Count: 96518, Percentage: 1.29%
- `a`: Count: 97085, Percentage: 1.29%
- `b`: Count: 97006, Percentage: 1.29%
- `c`: Count: 96269, Percentage: 1.28%
- `d`: Count: 96465, Percentage: 1.29%
- `e`: Count: 96326, Percentage: 1.28%
- `f`: Count: 96513, Percentage: 1.29%
- `g`: Count: 96040, Percentage: 1.28%
- `h`: Count: 96369, Percentage: 1.28%
- `i`: Count: 95803, Percentage: 1.28%
- `j`: Count: 95803, Percentage: 1.28%
- `k`: Count: 95580, Percentage: 1.27%
- `l`: Count: 95695, Percentage: 1.28%
- `m`: Count: 96105, Percentage: 1.28%
- `n`: Count: 95586, Percentage: 1.27%
- `o`: Count: 96450, Percentage: 1.29%
- `p`: Count: 96160, Percentage: 1.28%
- `q`: Count: 96170, Percentage: 1.28%
- `r`: Count: 95735, Percentage: 1.28%
- `s`: Count: 95711, Percentage: 1.28%
- `t`: Count: 95781, Percentage: 1.28%
- `u`: Count: 95751, Percentage: 1.28%
- `v`: Count: 96108, Percentage: 1.28%
- `w`: Count: 96668, Percentage: 1.29%
- `x`: Count: 95721, Percentage: 1.28%
- `y`: Count: 96190, Percentage: 1.28%
- `z`: Count: 96253, Percentage: 1.28%

</details>

<details>
<summary>Fixed implementation standard deviation: <strong>0.00%</strong></summary>

- `0`: Count: 121027, Percentage: 1.61%
- `1`: Count: 121161, Percentage: 1.62%
- `2`: Count: 120685, Percentage: 1.61%
- `3`: Count: 120983, Percentage: 1.61%
- `4`: Count: 120443, Percentage: 1.61%
- `5`: Count: 121489, Percentage: 1.62%
- `6`: Count: 120828, Percentage: 1.61%
- `7`: Count: 120678, Percentage: 1.61%
- `8`: Count: 120849, Percentage: 1.61%
- `9`: Count: 121047, Percentage: 1.61%
- `A`: Count: 120948, Percentage: 1.61%
- `B`: Count: 121242, Percentage: 1.62%
- `C`: Count: 120510, Percentage: 1.61%
- `D`: Count: 120392, Percentage: 1.61%
- `E`: Count: 120842, Percentage: 1.61%
- `F`: Count: 121073, Percentage: 1.61%
- `G`: Count: 120771, Percentage: 1.61%
- `H`: Count: 121041, Percentage: 1.61%
- `I`: Count: 120502, Percentage: 1.61%
- `J`: Count: 121002, Percentage: 1.61%
- `K`: Count: 121703, Percentage: 1.62%
- `L`: Count: 121166, Percentage: 1.62%
- `M`: Count: 120842, Percentage: 1.61%
- `N`: Count: 120937, Percentage: 1.61%
- `O`: Count: 121151, Percentage: 1.62%
- `P`: Count: 121720, Percentage: 1.62%
- `Q`: Count: 121288, Percentage: 1.62%
- `R`: Count: 121054, Percentage: 1.61%
- `S`: Count: 120891, Percentage: 1.61%
- `T`: Count: 120285, Percentage: 1.60%
- `U`: Count: 120922, Percentage: 1.61%
- `V`: Count: 121077, Percentage: 1.61%
- `W`: Count: 120977, Percentage: 1.61%
- `X`: Count: 120842, Percentage: 1.61%
- `Y`: Count: 121429, Percentage: 1.62%
- `Z`: Count: 120356, Percentage: 1.60%
- `a`: Count: 121471, Percentage: 1.62%
- `b`: Count: 121273, Percentage: 1.62%
- `c`: Count: 120997, Percentage: 1.61%
- `d`: Count: 121624, Percentage: 1.62%
- `e`: Count: 121058, Percentage: 1.61%
- `f`: Count: 120703, Percentage: 1.61%
- `g`: Count: 120633, Percentage: 1.61%
- `h`: Count: 120686, Percentage: 1.61%
- `i`: Count: 120924, Percentage: 1.61%
- `j`: Count: 120675, Percentage: 1.61%
- `k`: Count: 120682, Percentage: 1.61%
- `l`: Count: 120708, Percentage: 1.61%
- `m`: Count: 120933, Percentage: 1.61%
- `n`: Count: 121381, Percentage: 1.62%
- `o`: Count: 121289, Percentage: 1.62%
- `p`: Count: 121003, Percentage: 1.61%
- `q`: Count: 121037, Percentage: 1.61%
- `r`: Count: 120738, Percentage: 1.61%
- `s`: Count: 121067, Percentage: 1.61%
- `t`: Count: 121454, Percentage: 1.62%
- `u`: Count: 121116, Percentage: 1.61%
- `v`: Count: 121057, Percentage: 1.61%
- `w`: Count: 120582, Percentage: 1.61%
- `x`: Count: 120776, Percentage: 1.61%
- `y`: Count: 120843, Percentage: 1.61%
- `z`: Count: 121137, Percentage: 1.62%

</details>

#### Alphabetical Pattern: `[A-Za-z]{15}` _(500k iterations)_

<details>
<summary>Current Implementation standard deviation: <strong>0.01%</strong></summary>

- `A`: Count: 144695, Percentage: 1.93%
- `B`: Count: 144030, Percentage: 1.92%
- `C`: Count: 144560, Percentage: 1.93%
- `D`: Count: 144576, Percentage: 1.93%
- `E`: Count: 144199, Percentage: 1.92%
- `F`: Count: 144718, Percentage: 1.93%
- `G`: Count: 144377, Percentage: 1.93%
- `H`: Count: 144006, Percentage: 1.92%
- `I`: Count: 144809, Percentage: 1.93%
- `J`: Count: 144749, Percentage: 1.93%
- `K`: Count: 144114, Percentage: 1.92%
- `L`: Count: 143668, Percentage: 1.92%
- `M`: Count: 144593, Percentage: 1.93%
- `N`: Count: 144020, Percentage: 1.92%
- `O`: Count: 144532, Percentage: 1.93%
- `P`: Count: 144271, Percentage: 1.92%
- `Q`: Count: 144359, Percentage: 1.92%
- `R`: Count: 144024, Percentage: 1.92%
- `S`: Count: 144434, Percentage: 1.93%
- `T`: Count: 144253, Percentage: 1.92%
- `U`: Count: 144454, Percentage: 1.93%
- `V`: Count: 143844, Percentage: 1.92%
- `W`: Count: 144178, Percentage: 1.92%
- `X`: Count: 143611, Percentage: 1.91%
- `Y`: Count: 145117, Percentage: 1.93%
- `Z`: Count: 144671, Percentage: 1.93%
- `a`: Count: 143368, Percentage: 1.91%
- `b`: Count: 144101, Percentage: 1.92%
- `c`: Count: 143935, Percentage: 1.92%
- `d`: Count: 143822, Percentage: 1.92%
- `e`: Count: 143896, Percentage: 1.92%
- `f`: Count: 144278, Percentage: 1.92%
- `g`: Count: 143964, Percentage: 1.92%
- `h`: Count: 143997, Percentage: 1.92%
- `i`: Count: 144308, Percentage: 1.92%
- `j`: Count: 143973, Percentage: 1.92%
- `k`: Count: 144274, Percentage: 1.92%
- `l`: Count: 143926, Percentage: 1.92%
- `m`: Count: 144137, Percentage: 1.92%
- `n`: Count: 144251, Percentage: 1.92%
- `o`: Count: 145050, Percentage: 1.93%
- `p`: Count: 144199, Percentage: 1.92%
- `q`: Count: 144545, Percentage: 1.93%
- `r`: Count: 144659, Percentage: 1.93%
- `s`: Count: 143799, Percentage: 1.92%
- `t`: Count: 144169, Percentage: 1.92%
- `u`: Count: 143897, Percentage: 1.92%
- `v`: Count: 143322, Percentage: 1.91%
- `w`: Count: 144485, Percentage: 1.93%
- `x`: Count: 144276, Percentage: 1.92%
- `y`: Count: 144464, Percentage: 1.93%
- `z`: Count: 144043, Percentage: 1.92%

</details>

<details>
<summary>Fixed implementation standard deviation: <strong>0.00%</strong></summary>

- `A`: Count: 143610, Percentage: 1.91%
- `B`: Count: 144294, Percentage: 1.92%
- `C`: Count: 144220, Percentage: 1.92%
- `D`: Count: 144104, Percentage: 1.92%
- `E`: Count: 143579, Percentage: 1.91%
- `F`: Count: 144273, Percentage: 1.92%
- `G`: Count: 144645, Percentage: 1.93%
- `H`: Count: 143674, Percentage: 1.92%
- `I`: Count: 145014, Percentage: 1.93%
- `J`: Count: 143832, Percentage: 1.92%
- `K`: Count: 144136, Percentage: 1.92%
- `L`: Count: 144391, Percentage: 1.93%
- `M`: Count: 144045, Percentage: 1.92%
- `N`: Count: 144434, Percentage: 1.93%
- `O`: Count: 144567, Percentage: 1.93%
- `P`: Count: 143749, Percentage: 1.92%
- `Q`: Count: 144310, Percentage: 1.92%
- `R`: Count: 144527, Percentage: 1.93%
- `S`: Count: 144148, Percentage: 1.92%
- `T`: Count: 144304, Percentage: 1.92%
- `U`: Count: 144688, Percentage: 1.93%
- `V`: Count: 144368, Percentage: 1.92%
- `W`: Count: 143685, Percentage: 1.92%
- `X`: Count: 144036, Percentage: 1.92%
- `Y`: Count: 143993, Percentage: 1.92%
- `Z`: Count: 144601, Percentage: 1.93%
- `a`: Count: 144385, Percentage: 1.93%
- `b`: Count: 144710, Percentage: 1.93%
- `c`: Count: 144690, Percentage: 1.93%
- `d`: Count: 144621, Percentage: 1.93%
- `e`: Count: 144666, Percentage: 1.93%
- `f`: Count: 144058, Percentage: 1.92%
- `g`: Count: 144455, Percentage: 1.93%
- `h`: Count: 143737, Percentage: 1.92%
- `i`: Count: 144033, Percentage: 1.92%
- `j`: Count: 144028, Percentage: 1.92%
- `k`: Count: 143897, Percentage: 1.92%
- `l`: Count: 144746, Percentage: 1.93%
- `m`: Count: 143902, Percentage: 1.92%
- `n`: Count: 144264, Percentage: 1.92%
- `o`: Count: 144079, Percentage: 1.92%
- `p`: Count: 143746, Percentage: 1.92%
- `q`: Count: 143955, Percentage: 1.92%
- `r`: Count: 144321, Percentage: 1.92%
- `s`: Count: 144092, Percentage: 1.92%
- `t`: Count: 144343, Percentage: 1.92%
- `u`: Count: 144606, Percentage: 1.93%
- `v`: Count: 144434, Percentage: 1.93%
- `w`: Count: 144350, Percentage: 1.92%
- `x`: Count: 144374, Percentage: 1.92%
- `y`: Count: 144363, Percentage: 1.92%
- `z`: Count: 143918, Percentage: 1.92%

</details>

#### Numerical Pattern: `[0-9]{15}` _(500k iterations)_

<details>
<summary>Current Implementation standard deviation: <strong>0.01%</strong></summary>

- `0`: Count: 749682, Percentage: 10.00%
- `1`: Count: 749813, Percentage: 10.00%
- `2`: Count: 751089, Percentage: 10.01%
…