@yerTools 2025-06-20
Sorry, I didn't mean to be rude. 😅

I probably shouldn't have written a wall of text since the change is more or less trivial.

I don't think it's *that* relevant for security, but if a user creates a column in the admin interface with an autogenerated pattern, they probably expect more randomness than the current implementation, and this PR should fix it. Since you never know how certain features are (mis-)used.

Since those autogenerated patterns are also used for the row ID of a table, the entropy is lower than probably expected and collisions are more likely.

Anyway, sorry for this huge PR description. I hope the code changes speak for themselves then, and thank you very much not only for your answer but also for this amazing project! ❤️