@ganigeorgiev 2025-06-20
I was able to read through the PR description and I think I understand now what is the issue you are trying to solve.

I appreciate the tests and I'll run them tomorrow on a clear head, but at quick glance over the suggested implementation it feels a little error prone and difficult to read. Wouldn't it be easier to:
1. fill a slice with all runes from the pairs
2. generate a random number within the resulting length
3. return allRunes[random]

Or is performance the reason that you chose the current implementation?