@yerTools 2025-06-20
TBH I just tried to stick to the original code as much as possible and didn't think much about performance. 😇

I think you have a good point regarding readability.  
If you calculate the total number of runes to reduce allocations and GC pressure, most of the code will be the same.

---

<details>
<summary>You can take a look at the code for <code>allRunes</code> (with pre-allocation)</summary>

```go
func randomRuneFromPairs(pairs []rune) (rune, error) {
	if len(pairs)%2 != 0 {
		return 0, fmt.Errorf("invalid pairs slice: odd number of elements")
	}

	// Pre-calculate the total number of runes in all ranges.
	totalRunes := 0
	for i := 0; i < len(pairs); i += 2 {
		start, end := pairs[i], pairs[i+1]
		if start > end {
			return 0, fmt.Errorf("invalid range: start '%c' > end '%c'", start, end)
		}
		totalRunes += int(end - start + 1)
	}

	if totalRunes == 0 {
		return 0, errors.New("no runes to choose from")
	}

	// Create a slice to hold all runes in the ranges.
	allRunes := make([]rune, 0, totalRunes)
	for i := 0; i < len(pairs); i += 2 {
		start, end := pairs[i], pairs[i+1]
		for r := start; r <= end; r++ {
			allRunes = append(allRunes, r)
		}
	}

	// Select a random number in the range of total runes.
	runeNumber, err := randomNumber(totalRunes)
	if err != nil {
		return 0, fmt.Errorf("failed to generate random number: %w", err)
	}

	return allRunes[runeNumber], nil
}
```

</details>

---

<details>
<summary>You can take a look at the code for <code>allRunes</code> (without pre-allocation)</summary>

```go
func randomRuneFromPairs(pairs []rune) (rune, error) {
	if len(pairs)%2 != 0 {
		return 0, fmt.Errorf("invalid pairs slice: odd number of elements")
	}

	// Create a slice to hold all runes in the ranges.
	allRunes := make([]rune, 0)
	for i := 0; i < len(pairs); i += 2 {
		start, end := pairs[i], pairs[i+1]
		for r := start; r <= end; r++ {
			allRunes = append(allRunes, r)
		}
	}

	if len(allRunes) == 0 {
		return 0, errors.New("no runes to choose from")
	}

	// Select a random number in the range of total runes.
	runeNumber, err := randomNumber(len(allRunes))
	if err != nil {
		return 0, fmt.Errorf("failed to generate random number: %w", err)
	}

	return allRunes[runeNumber], nil
}
```

</details>

---

<details>
<summary>
The benchmark with pre-allocation produced similar results to the implementation before this PR.
But without pre-allocation the garbage collector has to work a lot more.
</summary>

| Implementation                           | Pattern              | Iterations | Time/op  | Memory/op | Allocs/op |
| ---------------------------------------- | -------------------- | ---------- | -------- | --------- | --------- |
| **Current (Buggy)**                      | `[a-z0-9]{15}`       | 369,079    | 3,395 ns | 1,840 B   | 97        |
| **Fixed (Previous)**                     | `[a-z0-9]{15}`       | 489,627    | 2,477 ns | 1,360 B   | 67        |
| **Fixed (`allRunes` pre-allocated)**     | `[a-z0-9]{15}`       | 344,450    | 3,206 ns | 3,280 B   | 67        |
| **Fixed (`allRunes` no pre-allocation)** | `[a-z0-9]{15}`       | 217,016    | 5,368 ns | 8,680 B   | 142       |
| ---                                      | ---                  | ---        | ---      | ---       | ---       |
| **Current (Buggy)**                      | `[A-Za-z0-9]{15}`    | 366,669    | 3,533 ns | 1,872 B   | 98        |
| **Fixed (Previous)**                     | `[A-Za-z0-9]{15}`    | 515,947    | 2,230 ns | 1,512 B   | 68        |
| **Fixed (`allRunes` pre-allocated)**     | `[A-Za-z0-9]{15}`    | 372,850    | 3,429 ns | 4,992 B   | 68        |
| **Fixed (`allRunes` no pre-allocation)** | `[A-Za-z0-9]{15}`    | 222,608    | 5,315 ns | 8,712 B   | 143       |
| ---                                      | ---                  | ---        | ---      | ---       | ---       |
| **Current (Buggy)**                      | `[A-Za-z0-9\-_]{30}` | 180,786    | 6,559 ns | 3,312 B   | 178       |
| **Fixed (Previous)**                     | `[A-Za-z0-9\-_]{30}` | 273,751    | 4,314 ns | 3,408 B   | 130       |
| **Fixed (`allRunes` pre-allocated)**     | `[A-Za-z0-9\-_]{30}` | 175,682    | 6,647 ns | 9,648 B   | 130       |
| **Fixed (`allRunes` no pre-allocation)** | `[A-Za-z0-9\-_]{30}` | 123,751    | 9,999 ns | 17,088 B  | 280       |

</details>

---

## Personal opinion - IMHO:

I think `Time/op` is not very important in this case.

As far as I'm concerned, this method is only used for auto-generated row values.
This means (IMHO) the following:

- Even with 10 µs per execution (on my cheap laptop with an `AMD Ryzen 5 8640HS`), this means there could be about 100,000 executions per second, and this call can also be run concurrently. I think in this case, when you have 100k SQLite inserts per second, this shouldn't be a bottleneck.

I think that readability and maintainability are way more important than performance, and there also isn't much room for performance improvements (at least in this PR).

---

If you like another implementation more than the one currently suggested by this PR, I can update it. 😄

Anyway, thank you very much! ❤️