@ganigeorgiev 2025-06-21
Thank you for trying it and rerunning the tests. I've also run them locally and the above reported numbers seems accurate.

I find the allRunes version more readable but considering that this is used only here and probably will never have to change, there is no need to bikeshed further so let's merge the PR as it is for now.

I'll try to prepare a minor release sometime later this weekend.